Fix: trapped without a trap, take two.
Been chasing this bug for days now, I think I finally figured it out.
It was always happening now with neothelids, which is (among other
things) an engulfer. Two rules about trapped monsters is 1) monsters
that are trapped cannot be engulfed, and 2) if an engulfer is trapped,
they can't use their engulfing attack. But number 2 was being bypassed,
and it was standing out with neothelids because:

- They can engulf (50% chance to switch to engulf attack)
- They spawn frequently in mind flayer nurseries
- The invalid engulf attempts while trapped led to inconsistent state

To zero in on the bug, I set up a condition where mind flayer larva
nurseries always spawn on every level, populate every available regular
room, and the odds of a neothelid appearing was always 100%. I also
relegated the fuzzer to staying on the regular dungeon. It didn't take
long to generate a ton of 'trapped without a trap' crash logs.

The fix: added a check in engulf_target() to prevent trapped monsters
from engulfing as a counter-attack.

There's another fix/tweak in here as well concerning mtrapped state and
worm segments.

Leaving the test code in this commit because after finding and fixing
the cause of the 'trapped without a trap' issue, I started seeing a lot
crashes conerning mind flayer larva as they attach themselves to a
humanoid and transform themselves into a mind flayer. So do not use this
code in production for a real game. You won't last long 😁