Fix: set hard upper limit for sparkle animation.
Quite a ways back, a patch was added to EvilHack to allow players to
manually set the sparkle animation to whatever value they wanted. By
default the value is 21, many players like to set this to a much lower
value, especially if they're playing on a system that isn't very fast.
The code protects against a negative value being set, but there was no
clamp on an upper limit. Sometimes the fuzzer would change the sparkle
animation value to something ridiculously high - watch it lag and take
forever to get to the next frame. This commits sets an upper limit to
match the default value of 21.