Fix: prevent various player races from using bane artrifacts.
It was noticed a few weeks ago that a player playing as vampire race had
permanently charged their alignment to lawful via altar sacrifice, and
then found the Mitre of Holiness in a bones pile, and was able to wear
and use the Mitre. This should not be possible, as the Mitre is bane
object vs undead. Turns out this restriction only worked if the player
polymorphed into an undead monster and then tried to use it, not if they
started out as undead.

The fix - add specific conditions in bane_applies() to handle these
scenarios. There was already an entry in here for Drow players trying to
wield artifacts that give off light. Fairly simple to expand that to
other races for other artifacts. Undead are covered for the Mitre as
well as Sunsword and the Hammer of the Gods, and demons for Demonbane
and the HotG.

Could probably add more here, but this is a good (and obvious) start.