Fix: prevent the spell 'change metal to wood' reverting objects back to their base material.
The issue: player tries to change a set of mithril dwarvish chain mail
into wood. Dwarvish chain mail can never normally be wood, so the
attemped failed, but the armor material reverted back to its base
material (iron), and on top of that, it's weight wasn't recalculated
after the fact.

This commit prevents any of that from happeneing. Did a bit of code
reorganization here as well.