Staff of the Archmagi affiliation change.
Prior to this commit, the Staff of the Archmagi was affiliated with
players of neutral alignment. Ok, but wizards that weren't neutral would
get blasted by wielding it. Which makes no sense really.

The SotA is now alignment agnostic, and affiliated with Wizards instead.
Any role can still use it (alignment doesn't matter), but unless you're
a Wizard, you'll be blasted by it. As you should be 😁