Fix: spellcasting monsters will gain experience killing other monsters with their spell attacks, take two.
This reverts commit c402363 - grow_up() is called in the AT_MAGC case in
mhitm.c instead of having to add multiple entries for it in mcastu.c

While looking at this, it seems that monsters never gained experience
from spit or breath attacks either, so that will be addressed in the
next commit. Vanilla NetHack never had monsters attack other monsters
with their special attack types, so it's no wonder that it's missing
here.