Grass and sand behave just like room tiles do when unseen, blocked, or in darkness.
Player have repeatedly mentioned that it felt unnatural that they could
see grass or sand regardless if those tiles were in a dark room, or
blocked by a boulder, etc. We'll give this change a try and see howq it
goes over.

I left ice tiles alone, as they are not affected by darkness/being
blocked in vanilla NetHack.