Changes to the scroll of consecration.
Having the quest nemesis spawn with a sack and then having it place the
scroll of consecration into it wasn't one of my best ideas. Especially
since not all quest nemesis can even use a bag (Ixoth comes to mind).
The reasoning for all this was so the scroll would be protected from the
elements. I have a better idea.

No more sacks for quest nemesis. The scroll is still in their inventory,
but now the scroll of consecration is fire/water/cancel proof. It is NOT
completely indestructible however - wielding it and whacking a black
dragon with it would not be a good idea. But it will resist the more
common effects that it could be subject to.