include crash-preventing 3.7 fix in 3.6
Using 'O' to interactively set a text match highlight for hunger can
trigger a crash.  Bug report #S14367 indicated in was an X11 issue
but it is more general than that.

Fixed a year ago with commit 4e2494c8595fce30de9a028a6ac0e21197c30972
in 3.7 when the fuzzer found it.