Fix: migrations of covetous monsters.
So this has been a long-standing issue, had been reported a handful of
times over the last couple of years. The last report was in the form of
a reddit post (see https://redd.it/1nvlfk3/) where the player's pet
arch-lich fell down a trap door on the first level of Goblin Town. They
should have fallen down to Gollum's lair below, but instead wound up in
Fort Ludios.

What was happening? In a nutshell, the tactics() function return value
was being ignored for covetous monsters within dochug() and would
continue to process as if the monster were still on the current level.
set_apparxy() would be called later on in the chain and would clobber
the monster migration destination values, using the player's x,y
coordinates instead. In this specific scenario, the player was standing
on position (5,1), which translated to dnum 5, dlevel 1. dnum (dungeon
number) is Fort Ludios, dlevel (dungeon level) was 1.

Community member Furey did all the leg work, tracing the code paths,
figuring out the root cause of the bug. Made it fairly easy for me to
develop a fix. Basically, don't void the return value of tactics(), and
depending on its value, stop processing, or continue (if the monster
died or left the level, stop processing).