Fix: placing lit objects into a container while blind would not extinguish the object.
Another one of those cases where something can or can't happen based on
whether you can see it or not. The fix was to simply decouple the action
from visibility of the action. Having lit objects within a container can
lead to other issues, so good that this was identified.