Fix: dungeons overview and grass/sand/trees.
If the player was only aware of one grass or one sand tile and then used
dungeon overview, the terrain reported was grammatically incorrect ('a
sand' or 'a grass'). Trees/dead trees had the same problem, but only if
two trees were known ('some tree' or 'some dead tree'). Corrected this
by creating a new macro specifically for mass nouns (a noun that is
uncountable) and having it adjust grass and sand accordingly.