Fix: negative damage applied to monsters under certain conditions.
The function exclam() returns a ? instead of ! if force (amount of
damage) is less than zero. Some players had reported seeing monsters
being hit and the feedback ending in ? and didn't know what to make of
it.

A ways back, I had added more conditions to weapon.c for damage bonuses
or penalties depending on the type of weapons used and what the monster
is made of. Case in point: skeletons and other skeletal monsters get a
damage reduction with attacked with slashing or piercing weapons. This
damage penalty wasn't properly clamped to not go below zero, hence the ?
at the end of various hit messages. This commit fixes this oversight.