Fix: entangle spell properly entangles the player.
Now that spellcasting monsters can learn and cast the spell 'entangle',
it was time to have it properly ensnare the player.

Created a new trap condition, TT_ENTANGLED, which will immobilize the
entangled player, but they can still take action. Unlike being caught in
a bear trap, players can't pull free, they'll have to wait for the spell
to wear off (3-8 turns if cast on oneself or if entangled by a monster
casting the spell).