Fix: revamp how spellcasting monsters use the player spell charm monster.
The initial implementation for this wasn't very good. The spellcasting
monster would cast this at the player just like any other
immediate-based spell instead of at a tame pet. This has now been
reworked so the spellcasting monster will cast this directly at a pet it
wants to actually attack (being attacked by your pet or it has a grudge
versus it). If the player's pet doesn't engage, they won't be targeted.

Charm monster is a high level spell in EvilHack, so this spell would be
off-limits to most spellcasting monsters. But I think this so-called
'flipping the script' of the spell that actually untames your pet when
cast by a hostile monster is awesome - the spellcasting monster just
'tamed' your pet and made it its own. Decidedly an EvilHack thing to do
😁