Fix: refactor mchoose_learned_spell() to accept coordinates.
Reason: so a spellcasting monster that has learned the player spell
'entangle' will intelligently determine if the spell would be effective.
Entangle only works if the target is standing on grass or near a live
tree. The function cast_learned_spell() has been restructured to 1)
determine the targets coordinates first, 2) pass them to
mchoose_learned_spell() so the spell entangle is excluded when there's
no vegetation, and 3) the monster will then pick a different learned
spell if entangle can't work. No wasted spell attack. This also sets the
stage for other spells that may need consideration depending on where
the target is located or what terrain they're on or near.