Add remaining NODIR-based spells for monster learned spellcasting.
Those being: jumping, levitation, repair armor, and remove curse.
Jumping and levitation are on a timer and currently last for 50-100
turns.

Also added some logic for the healing-based spells - if the spellcasting
monster is blind or ill and they are of a high enough level to cast the
spells, cure blindness/sickness will take priority.