Add burning hands and shocking grasp spells for monster learned spellcasting.
These are also NODIR type spells, as the spellcasting monster can only
cast them on themselves. These behave pretty much the same way that they
do for the player.

These were the last two spells on the list to add, for a total of 26
player spells that spellcasting monsters can learn.