Develop an alternate quest for Dark (chaotic) Knight, round two.
In this commit:
* Create the actual quest leader/nemesis - Morgan le Fay as the quest
  leader, Merlin as the quest nemesis. Both are powerful spellcasters
* Change the names of their locations that can be found in Arthurian
  lore - the Isle of Avalon (Morgan le Fay) and the Caledonian Forest
  (Merlin)
* Depending on the lore you read, Merlin was considered a Druid of
  sorts, and the storyline to be developed will focus on that (woodland
  home, befriended by woodland creatures, etc)

Next up, the quest text so it matches a legit storyline, and not the
placeholder text that was added initially.

Did you know that Merlin was half-demon? But chose to be good
obviously... not sure if I can flag him as MH_DEMON as well as MH_HUMAN
without introducing behavior that doesn't belong here.