Fix: special handling when wishing for the Magic Mirror of Merlin and which owner shows up.
Because Merlin is now an actual monster within the game, special
handling when parsing the artifact 'magic mirror of merlin' needed to be
added. Also, since there are now two different quest leaders for the
Knight's quest depending on what alignment you choose, another special
case needed to be made to determine which owner shows up when the MMoM
is wished for.