More enhanced livelogging of unique monsters.
Right after the initial 'enhanced livelogging' commit, I realized there
wouldn't be a livelog at all if a unique monster died by other means not
due to the player or their pet. So now if a unique monster dies, say
from a grudge attack from a bunch of zombies that aren't pets, the
following livelog would be something like 'The Goblin King has been
killed'. People seeing this will know it's been killed, but not due to
the player or their pet.