Fix: enhanced livelogging output.
After looking at the changes with enhanced livelogging, it became clear
that some of the out would not fit. Tweaked the livelogging when a
player's pet kills a unique monster, and had to also make an edit to the
EvilHack channel bot as well so the output would be grammatically. Also
decided to ditch the unique monster livelogging if they weren't killed
by the player or their pet, it did not fit really at all.