Honey badgers prioritize royal jelly over the player.
This was the original intent long ago when honey badgers were
introduced, this commit makes it a reality. If honey badgers spawn on
the same level as a beehive, instead of seeking out the player like the
vast majority of monsters do, they will instead seek out the royal jelly
in the beehive first. And honey badgers love to eat royal jelly. Who can
reach the beehive first?